requirements.txt
# Python dependencies for the EIR system

# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
openlm>=0.0.5
python-dotenv>=1.0.0

# Optional dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0